+++ packages/app-contracts/src/Codes/Code.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { RouteComponentProps } from 'react-router';
 import { withRouter } from 'react-router-dom';
-import { Button, Card, CodeSummary, Forget, Icon } from '@polkadot/ui-app';
+import { Button, Card, CodeRow, Forget, Icon } from '@polkadot/ui-app';
 import { withMulti } from '@polkadot/ui-api';
 
 import ABI from '../ABI';
@@ -47,7 +47,7 @@ class Contract extends React.PureComponent<Props> {
     return (
       <CodeCard>
         {this.renderModals()}
-        <CodeSummary
+        <CodeRow
           buttons={this.renderButtons()}
           code={code}
           isEditable
@@ -58,7 +58,7 @@ class Contract extends React.PureComponent<Props> {
             onChange={this.onChangeABI}
             onRemove={this.toggleRemoveABI}
           />
-        </CodeSummary>
+        </CodeRow>
       </CodeCard>
     );
   }
