+++ packages/app-contracts/src/RemoveABI.tsx
@@ -6,7 +6,7 @@ import { I18nProps } from '@polkadot/ui-app/types';
 import { CodeStored } from '@polkadot/app-contracts/types';
 
 import React from 'react';
-import { Button, CodeSummary, Modal } from '@polkadot/ui-app';
+import { Button, CodeRow, Modal } from '@polkadot/ui-app';
 
 import translate from './translate';
 
@@ -74,11 +74,11 @@ class RemoveABI extends React.PureComponent<Props> {
     const { code } = this.props;
 
     return (
-      <CodeSummary
+      <CodeRow
         code={code!}
       >
         {this.content()}
-      </CodeSummary>
+      </CodeRow>
     );
   }
