+++ packages/ui-app/src/index.tsx
@@ -18,7 +18,7 @@ export { default as CardSummary } from './CardSummary';
 export { default as Chart } from './Chart';
 export { default as Columar } from './Columar';
 export { default as Column } from './Column';
-export { default as CodeSummary } from './CodeSummary';
+export { default as CodeRow } from './CodeRow';
 export { default as CopyButton } from './CopyButton';
 export { default as CryptoType } from './CryptoType';
 export { default as Dropdown } from './Dropdown';
